+++ lib/src/config.dart
@@ -4,6 +4,13 @@ import 'package:emoji_picker_flutter/src/category_icons.dart';
 import 'package:emoji_picker_flutter/src/emoji_picker.dart';
 import 'package:flutter/material.dart';
 
+/// Default Widget if no recent is available
+const DefaultNoRecentsWidget = Text(
+  'No Recents',
+  style: TextStyle(fontSize: 20, color: Colors.black26),
+  textAlign: TextAlign.center,
+);
+
 /// Config for customizations
 class Config {
   /// Constructor
@@ -24,11 +31,7 @@ class Config {
       this.enableSkinTones = true,
       this.showRecentsTab = true,
       this.recentsLimit = 28,
-      this.noRecentsText = const Text(
-        'No Recents',
-        style: TextStyle(fontSize: 20, color: Colors.black26),
-        textAlign: TextAlign.center,
-      ),
+      this.noRecents = DefaultNoRecentsWidget,
       this.tabIndicatorAnimDuration = kTabScrollDuration,
       this.categoryIcons = const CategoryIcons(),
       this.buttonMode = ButtonMode.MATERIAL});
@@ -84,7 +87,7 @@ class Config {
   final int recentsLimit;
 
   /// A widget (usually [Text]) to be displayed if no recent emojis to display
-  final Widget noRecentsText;
+  final Widget noRecents;
 
   /// Duration of tab indicator to animate to next category
   final Duration tabIndicatorAnimDuration;
@@ -146,7 +149,7 @@ class Config {
         other.enableSkinTones == enableSkinTones &&
         other.showRecentsTab == showRecentsTab &&
         other.recentsLimit == recentsLimit &&
-        other.noRecentsText == noRecentsText &&
+        other.noRecents == noRecents &&
         other.tabIndicatorAnimDuration == tabIndicatorAnimDuration &&
         other.categoryIcons == categoryIcons &&
         other.buttonMode == buttonMode;
@@ -170,7 +173,7 @@ class Config {
       enableSkinTones.hashCode ^
       showRecentsTab.hashCode ^
       recentsLimit.hashCode ^
-      noRecentsText.hashCode ^
+      noRecents.hashCode ^
       tabIndicatorAnimDuration.hashCode ^
       categoryIcons.hashCode ^
       buttonMode.hashCode;